requirements.txt
pyswisseph==2.10.3.2
pytz==2024.1
mido==1.3.2  # for Phase 3
